+++ playground/nuxt.config.ts
@@ -6,4 +6,14 @@ export default defineNuxtConfig({
 
   devtools: { enabled: true },
   compatibilityDate: '2024-07-14',
+
+  nitro: {
+    prerender: {
+      failOnError: false,
+    },
+  },
+
+  scripts: {
+    debug: true,
+  },
 })