+++ playground/pages/third-parties/youtube/nuxt-scripts.vue
@@ -14,6 +14,9 @@ function changeVideoFallback() {
 
 <template>
   <div>
+    <NuxtLink to="/third-parties/youtube/multiple" class="block underline mb-5">
+      Multiple YouTube Players
+    </NuxtLink>
     <div>
       <ScriptYouTubePlayer :video-id="videoid" above-the-fold>
         <template #awaitingLoad>