+++ src/runtime/registry/intercom.ts
@@ -2,7 +2,7 @@ import { joinURL } from 'ufo'
 import { useRegistryScript } from '../utils'
 import { literal, number, object, optional, string, union } from '#nuxt-scripts-validator'
 import type { InferInput } from '#nuxt-scripts-validator'
-import type { RegistryScriptInput } from '#nuxt-scripts'
+import type { RegistryScriptInput } from '#nuxt-scripts/types'
 
 export const IntercomOptions = object({
   app_id: string(),
@@ -20,27 +20,27 @@ export type IntercomInput = RegistryScriptInput<typeof IntercomOptions, true, fa
 
 export interface IntercomApi {
   Intercom: ((event: 'boot', data?: InferInput<typeof IntercomOptions>) => void)
-  & ((event: 'shutdown') => void)
-  & ((event: 'update', options?: InferInput<typeof IntercomOptions>) => void)
-  & ((event: 'hide') => void)
-  & ((event: 'show') => void)
-  & ((event: 'showSpace', spaceName: 'home' | 'messages' | 'help' | 'news' | 'tasks' | 'tickets' | string) => void)
-  & ((event: 'showMessages') => void)
-  & ((event: 'showNewMessage', content?: string) => void)
-  & ((event: 'onHide', fn: () => void) => void)
-  & ((event: 'onShow', fn: () => void) => void)
-  & ((event: 'onUnreadCountChange', fn: () => void) => void)
-  & ((event: 'trackEvent', eventName: string, metadata?: Record<string, any>) => void)
-  & ((event: 'getVisitorId') => Promise<string>)
-  & ((event: 'startTour', tourId: string | number) => void)
-  & ((event: 'showArticle', articleId: string | number) => void)
-  & ((event: 'showNews', newsItemId: string | number) => void)
-  & ((event: 'startSurvey', surveyId: string | number) => void)
-  & ((event: 'startChecklist', checklistId: string | number) => void)
-  & ((event: 'showTicket', ticketId: string | number) => void)
-  & ((event: 'showConversation', conversationId: string | number) => void)
-  & ((event: 'onUserEmailSupplied', fn: () => void) => void)
-  & ((event: string, ...params: any[]) => void)
+    & ((event: 'shutdown') => void)
+    & ((event: 'update', options?: InferInput<typeof IntercomOptions>) => void)
+    & ((event: 'hide') => void)
+    & ((event: 'show') => void)
+    & ((event: 'showSpace', spaceName: 'home' | 'messages' | 'help' | 'news' | 'tasks' | 'tickets' | string) => void)
+    & ((event: 'showMessages') => void)
+    & ((event: 'showNewMessage', content?: string) => void)
+    & ((event: 'onHide', fn: () => void) => void)
+    & ((event: 'onShow', fn: () => void) => void)
+    & ((event: 'onUnreadCountChange', fn: () => void) => void)
+    & ((event: 'trackEvent', eventName: string, metadata?: Record<string, any>) => void)
+    & ((event: 'getVisitorId') => Promise<string>)
+    & ((event: 'startTour', tourId: string | number) => void)
+    & ((event: 'showArticle', articleId: string | number) => void)
+    & ((event: 'showNews', newsItemId: string | number) => void)
+    & ((event: 'startSurvey', surveyId: string | number) => void)
+    & ((event: 'startChecklist', checklistId: string | number) => void)
+    & ((event: 'showTicket', ticketId: string | number) => void)
+    & ((event: 'showConversation', conversationId: string | number) => void)
+    & ((event: 'onUserEmailSupplied', fn: () => void) => void)
+    & ((event: string, ...params: any[]) => void)
 }
 
 declare global {