+++ src/runtime/registry/x-pixel.ts
@@ -1,7 +1,7 @@
 import type { UseScriptInput } from '@unhead/vue'
 import { useRegistryScript } from '../utils'
 import { object, optional, string } from '#nuxt-scripts-validator'
-import type { RegistryScriptInput } from '#nuxt-scripts'
+import type { RegistryScriptInput } from '#nuxt-scripts/types'
 
 interface ContentProperties {
   content_type?: string | null