+++ src/runtime/utils.ts
@@ -1,15 +1,18 @@
 import { defu } from 'defu'
 import type { GenericSchema, InferInput, ObjectSchema, ValiError } from 'valibot'
 import type { UseScriptInput } from '@unhead/vue'
+import { useRuntimeConfig } from 'nuxt/app'
 import { useScript } from './composables/useScript'
 import { parse } from '#nuxt-scripts-validator'
-import { useRuntimeConfig } from '#imports'
 import type {
   EmptyOptionsSchema,
+  InferIfSchema,
   NuxtUseScriptOptions,
   RegistryScriptInput,
   ScriptRegistry,
-} from '#nuxt-scripts'
+} from '#nuxt-scripts/types'
+
+export type MaybePromise<T> = Promise<T> | T
 
 function validateScriptInputSchema<T extends GenericSchema>(key: string, schema: T, options?: InferInput<T>) {
   if (import.meta.dev) {
@@ -25,10 +28,10 @@ function validateScriptInputSchema<T extends GenericSchema>(key: string, schema:
   }
 }
 
-type OptionsFn<O extends ObjectSchema<any, any>> = (options: InferInput<O>) => ({
+type OptionsFn<O> = (options: InferIfSchema<O>) => ({
   scriptInput?: UseScriptInput
   scriptOptions?: NuxtUseScriptOptions
-  schema?: O
+  schema?: O extends ObjectSchema<any, any> ? O : undefined
   clientInit?: () => void
 })
 
@@ -36,10 +39,10 @@ export function scriptRuntimeConfig<T extends keyof ScriptRegistry>(key: T) {
   return ((useRuntimeConfig().public.scripts || {}) as ScriptRegistry)[key]
 }
 
-export function useRegistryScript<T extends Record<string | symbol, any>, O extends ObjectSchema<any, any> = EmptyOptionsSchema, U = {}>(registryKey: keyof ScriptRegistry | string, optionsFn: OptionsFn<O>, _userOptions?: RegistryScriptInput<O>) {
+export function useRegistryScript<T extends Record<string | symbol, any>, O = EmptyOptionsSchema, U = {}>(registryKey: keyof ScriptRegistry | string, optionsFn: OptionsFn<O>, _userOptions?: RegistryScriptInput<O>) {
   const scriptConfig = scriptRuntimeConfig(registryKey as keyof ScriptRegistry)
   const userOptions = Object.assign(_userOptions || {}, typeof scriptConfig === 'object' ? scriptConfig : {})
-  const options = optionsFn(userOptions)
+  const options = optionsFn(userOptions as InferIfSchema<O>)
 
   const scriptInput = defu(userOptions.scriptInput, options.scriptInput, { key: registryKey }) as any as UseScriptInput
   const scriptOptions = Object.assign(userOptions?.scriptOptions || {}, options.scriptOptions || {})
@@ -72,4 +75,14 @@ export function useRegistryScript<T extends Record<string | symbol, any>, O exte
     }
   }
   return useScript<T, U>(scriptInput, scriptOptions as NuxtUseScriptOptions<T, U>)
+}
+
+export function pick(obj: Record<string, any>, keys: string[]) {
+  const res: Record<string, any> = {}
+  for (const k of keys) {
+    if (k in obj) {
+      res[k] = obj[k]
+    }
+  }
+  return res
 }