+++ CHANGELOG.md
@@ -1,6 +1,196 @@
 # Changelog
 
 
+## v0.9.5
+
+[compare changes](https://github.com/nuxt/scripts/compare/v0.9.4...v0.9.5)
+
+### 🚀 Enhancements
+
+- **googleMaps:** Language & region params ([#286](https://github.com/nuxt/scripts/pull/286))
+- Improved script warmup ([#302](https://github.com/nuxt/scripts/pull/302))
+
+### 🩹 Fixes
+
+- Avoid warming delayed script src ([357d02a](https://github.com/nuxt/scripts/commit/357d02a))
+- **stripe:** Prefer `@stripe/stripe-js` over `@types/stripe-v3` ([#300](https://github.com/nuxt/scripts/pull/300))
+- Optional Valibot schema ([#287](https://github.com/nuxt/scripts/pull/287))
+
+### 📖 Documentation
+
+- Fix typo/incomplete sentence ([#285](https://github.com/nuxt/scripts/pull/285))
+
+### 🏡 Chore
+
+- Type issue ([6490ce3](https://github.com/nuxt/scripts/commit/6490ce3))
+- Type issue ([0c5135c](https://github.com/nuxt/scripts/commit/0c5135c))
+- Playground for multi datalayers ([#173](https://github.com/nuxt/scripts/pull/173))
+- Test scripts ([1f7df13](https://github.com/nuxt/scripts/commit/1f7df13))
+- Broken ci ([c89380f](https://github.com/nuxt/scripts/commit/c89380f))
+- **devtools:** Fix deprecation ([29864f1](https://github.com/nuxt/scripts/commit/29864f1))
+- Bump deps & lint ([c54412f](https://github.com/nuxt/scripts/commit/c54412f))
+
+### ❤️ Contributors
+
+- Harlan ([@harlan-zw](http://github.com/harlan-zw))
+- Harlan Wilton ([@harlan-zw](http://github.com/harlan-zw))
+- Julien Huang ([@huang-julien](http://github.com/huang-julien))
+- Morgan-retex ([@morgan-retex](http://github.com/morgan-retex))
+- Stefano Bartoletti ([@stefanobartoletti](http://github.com/stefanobartoletti))
+
+## v0.9.4
+
+[compare changes](https://github.com/nuxt/scripts/compare/v0.9.3...v0.9.4)
+
+### 🩹 Fixes
+
+- **matomo:** Easier cloud config using `cloudId` ([d7e18c4](https://github.com/nuxt/scripts/commit/d7e18c4))
+- **matomo:** Support numeric `siteId` ([299516c](https://github.com/nuxt/scripts/commit/299516c))
+- Avoid overriding `<link rek="preload"` with invalid src ([25f1ade](https://github.com/nuxt/scripts/commit/25f1ade))
+
+### 📖 Documentation
+
+- Fix docs for disabling scripts in development ([#281](https://github.com/nuxt/scripts/pull/281))
+
+### 🏡 Chore
+
+- Bump deps ([931f4f6](https://github.com/nuxt/scripts/commit/931f4f6))
+
+### ❤️ Contributors
+
+- Harlan ([@harlan-zw](http://github.com/harlan-zw))
+- Paul ([@schnetzi](http://github.com/schnetzi))
+
+## v0.9.3
+
+[compare changes](https://github.com/nuxt/scripts/compare/v0.9.2...v0.9.3)
+
+### 🩹 Fixes
+
+- Use free path for virtual plugin template ([8624923](https://github.com/nuxt/scripts/commit/8624923))
+- 🐛  Fixed "height" css style property ([#275](https://github.com/nuxt/scripts/pull/275))
+- **adsense:** Push ad slot once only on dev ([#276](https://github.com/nuxt/scripts/pull/276))
+
+### 📖 Documentation
+
+- Update scripts:registry hook example ([#271](https://github.com/nuxt/scripts/pull/271))
+- ✏️  Updated definition for YouTube Player ([#273](https://github.com/nuxt/scripts/pull/273))
+
+### 🏡 Chore
+
+- Bump deps ([81f0923](https://github.com/nuxt/scripts/commit/81f0923))
+
+### ❤️ Contributors
+
+- Harlan ([@harlan-zw](http://github.com/harlan-zw))
+- Asoka Wotulo ([@asokawotulo](http://github.com/asokawotulo))
+- Arb85 ([@arb85](http://github.com/arb85))
+- Bernhard Berger ([@bernhardberger](http://github.com/bernhardberger))
+
+## v0.9.2
+
+[compare changes](https://github.com/nuxt/scripts/compare/v0.9.1...v0.9.2)
+
+### 🩹 Fixes
+
+- Types for ref triggers ([7b3570d](https://github.com/nuxt/scripts/commit/7b3570d))
+- Cache bundled scripts relative to nuxt root ([cc5e01b](https://github.com/nuxt/scripts/commit/cc5e01b))
+
+### 🏡 Chore
+
+- Bump deps ([582d360](https://github.com/nuxt/scripts/commit/582d360))
+- Missing import ([9e33732](https://github.com/nuxt/scripts/commit/9e33732))
+
+### ❤️ Contributors
+
+- Harlan ([@harlan-zw](http://github.com/harlan-zw))
+
+## v0.9.1
+
+[compare changes](https://github.com/nuxt/scripts/compare/v0.9.0...v0.9.1)
+
+### 🩹 Fixes
+
+- Prefer ref triggers or promise booleans ([bacd3de](https://github.com/nuxt/scripts/commit/bacd3de))
+
+### 🏡 Chore
+
+- Ensure dir for script ([e4f140c](https://github.com/nuxt/scripts/commit/e4f140c))
+- Recommend 1.11.5 unhead ([9c1a3e3](https://github.com/nuxt/scripts/commit/9c1a3e3))
+
+### ❤️ Contributors
+
+- Harlan ([@harlan-zw](http://github.com/harlan-zw))
+
+## v0.9.0
+
+[compare changes](https://github.com/nuxt/scripts/compare/v0.8.5...v0.9.0)
+
+### 🚀 Enhancements
+
+- Allow fallback on bundle fail ([#255](https://github.com/nuxt/scripts/pull/255))
+
+### 🩹 Fixes
+
+- **adsense:** ⚠️  Remove default `data-ad-format` ([#248](https://github.com/nuxt/scripts/pull/248))
+- Improved `bundle: true` debug and cache clearing ([d269066](https://github.com/nuxt/scripts/commit/d269066))
+- **metaPixel:** Use meta logo instead of facebook logo ([#264](https://github.com/nuxt/scripts/pull/264))
+- Insert script src with `app.baseURL` ([52e1fcb](https://github.com/nuxt/scripts/commit/52e1fcb))
+
+### 🏡 Chore
+
+- Broken type checking ([97a5218](https://github.com/nuxt/scripts/commit/97a5218))
+- Bump deps and lint ([e73b4c6](https://github.com/nuxt/scripts/commit/e73b4c6))
+- Fix playground ([575c4dd](https://github.com/nuxt/scripts/commit/575c4dd))
+
+#### ⚠️ Breaking Changes
+
+- **adsense:** ⚠️  Remove default `data-ad-format` ([#248](https://github.com/nuxt/scripts/pull/248))
+
+### ❤️ Contributors
+
+- Harlan ([@harlan-zw](http://github.com/harlan-zw))
+- 2nofa11 ([@2nofa11](http://github.com/2nofa11))
+- Julien Huang ([@huang-julien](http://github.com/huang-julien))
+- Asoka Wotulo ([@asokawotulo](http://github.com/asokawotulo))
+
+## v0.8.5
+
+[compare changes](https://github.com/nuxt/scripts/compare/v0.8.4...v0.8.5)
+
+### 🩹 Fixes
+
+- **bundling:** Throw error if download fails ([#253](https://github.com/nuxt/scripts/pull/253))
+- Group failed scripts log ([#254](https://github.com/nuxt/scripts/pull/254))
+- **YouTube:** Broken api types ([a00f753](https://github.com/nuxt/scripts/commit/a00f753))
+
+### 🏡 Chore
+
+- Maybe fix vimeo types ([159fdcd](https://github.com/nuxt/scripts/commit/159fdcd))
+- Bump deps ([e0034fa](https://github.com/nuxt/scripts/commit/e0034fa))
+
+### ❤️ Contributors
+
+- Harlan <harlan@harlanzw.com>
+- Julien Huang <julien.h.dev@gmail.com>
+
+## v0.8.4
+
+[compare changes](https://github.com/nuxt/scripts/compare/v0.8.3...v0.8.4)
+
+### 🩹 Fixes
+
+- **ScriptGoogleMaps:** Clean up map markers properly ([53bb530](https://github.com/nuxt/scripts/commit/53bb530))
+
+### 🏡 Chore
+
+- Bump deps and lint ([06c757c](https://github.com/nuxt/scripts/commit/06c757c))
+- Sync lock ([bdc5f98](https://github.com/nuxt/scripts/commit/bdc5f98))
+
+### ❤️ Contributors
+
+- Harlan <harlan@harlanzw.com>
+
 ## v0.8.3
 
 [compare changes](https://github.com/nuxt/scripts/compare/v0.8.2...v0.8.3)